Shiny Hunt Position Swapping
Implemented a way to swap the positions of shiny hunts using move up/down arrows in the menu that opens when long clicking a shiny hunt. Swapped positions currently only persist for the lifespan of the main menu instance (i.e. the swapped positions are not being committed to the database yet).