Added Gen 8 Pokemon + Repositioned Form Name
Added all Gen 8 Pokemon/Forms into the database. In individual_hunt.xml and IndividualHunt.kt, separated the Pokemon name and form so that the form name is displayed on a separate line. Slightly reduced the size of the Pokemon image to compensate so that the "Select a Pokemon" button doesn't get cut off in landscape mode.