Implemented Clear Filters
Implemented the "Clear Filters" functionality from both the main page and the filter menu. Selecting to clear the filters produces a confirmation dialog. Confirming the filter clearing will automatically collapse all shiny hunts (to avoid the "Expand All" checkbox being checked while some hunts aren't expanded). Also added a setMoveButtonsEnabled() function in the ShinyHuntListAdapter so that the move buttons can be toggled from MainActivity. This allows the move buttons to be disabled when filters are currently being applied and enabled when filters are cleared (although the sort method still needs to be DEFAULT).