Added Game Filters + Better Dialog Handling
Implemented filters for origin and current games. Modified the GameSelectionAdapter so it can keep track of the currently selected games, highlighting them in the recycler view. Improved the handling of all modals in both MainActivity and IndividualHunt so that the user can't open multiple modals at the same time (via spamming the buttons).