Sort Button Revamp
Relocated the "floating" Sort button to the top toolbar so that it no longer blocks the last shiny hunt in the recycler view. Instead of opening a side menu, the Sort button now opens a dialog with a spinner to select the sort method, a radio group with two options to select the sort order (descending or ascending), and a "Confirm Sort" button to apply the selected sort options to the shiny hunt recycler view. Also enhanced the background of the filter and sort dialogs (i.e. applied the background resource to the entire dialog windows instead of just their views).