Fixed Bug with Unsaved Changes Detection
Found a bug where the back button would still detect unsaved changes even if the user didn't edit anything. Identified that the bug was the result of an oversight with how incomplete shiny hunts were being saved. When the completion checkbox is not checked while saving, the app would automatically set the hunt's finish date to null in the database, but the back button compares the text in the finish date field (which is an empty string by default) with the value of the finish date in the database. Thus, the back button detected an "unsaved change" as it was comparing an empty string to a null value. Modified the save button's on click listener so that it saves the finish date as an empty string for incomplete hunts.