Fixed bug with search bar in IndividualHunt Pokemon selection
The search bar in the IndividualHunt Pokemon selection dialog wasn't working because it was being accessed from a different layout ('selectPokemonDialog') than the one being used for creating the dialog ('selectPokemonDialogLayout'). To fix naming confusion, reorganized the code to use a single layout 'selectPokemonDialogLayout' for the Pokemon selection dialog, and used 'selectPokemonDialog' as the name of the actual dialog. To prevent errors when reopening the dialog, added a line to detach the dialog layout from any previous parent before reattaching the layout.