MyApplication + Counter Multiplier
Created an application file 'MyApplication.kt' to store globally accessible variables. Reconfigured 'AndroidManifest.xml' to use this application.

Added a "counter multiplier" button and dialog in 'MainActivity.kt'. A 'counterMultiplier' is stored globally in 'MyApplication.kt' and can be modified to adjust the amount that the increment/decrement buttons change the counters in the shiny hunt list.