Fixed bug with instantiation MyApplication + Added constant span counts for shiny hunt recycler view
Fixed a bug with the previous commit that was causing the app to crash ('MyApplication' was attempted to be instantiated in 'MainActivity.kt' outside of the 'onCreate()' function, but it can only be instantiated after 'onCreate()' calls 'super.onCreate()').

Added two more constants in 'MyApplication.kt' for the shiny hunt recycler view's span counts. Replaced the "magic" shiny hunt recycler view span counts in 'MainActivity.kt' with these constants.