Dynamic UI Sizing in the Main Page
Completely refactored activity_main.xml and shiny_hunt_item.xml to use ConstraintLayouts instead of LinearLayouts, allowing the top bar and the shiny hunt recycler view in the Main Page to automatically resize based on the dimensions of the device.

Replaced the integer constants (for span counts of pokemon, game, and shiny hunt recycler views) in MyApplication.kt with integer resources in the res directory. values/integers.xml contains the integer resources for phones, and values-sw600dp/integers.xml contains the integer resources for tablets.