Major UI Texture Upgrade + Pokemon Selection Refactoring/Debugging
Updated a lot of the drawable UI textures. Most notably, changed the designated "completed hunt" color to gold so that green can be used as the designated color for selection.

Updated the look of the Pokemon Filter dialog to display a transparent green overlay above selected Pokemon and Pokemon Forms (rather than in the background like before). The Selected Pokemon RecyclerView also uses a transparent background instead of the incomplete hunt gradient.

Refactored and debugged MainActivity and the Pokemon-related adapters so that selected Pokemon items are always collapsed by default to avoid cluttering the screen when the user selects Pokemon with a lot of forms. Also gave each selected Pokemon item a TextView displaying the number of its forms that are currently selected.