Pokemon Unselection Animation Fix + Constant Transition Duration
The use of notifyDataSetChanged() was causing the transition animation to not play when unselecting a Pokemon directly through the pokemonListRecyclerView or via its "unselect" button in the selectedPokemonRecyclerView. Replaced notifyDataSetChanged() in PokemonSelectionAdapter's updateSelectedPokemon() function with notifyItemChanged() for each position that has changed.

Replaced all hard-coded transition durations (Pokemon selection and Game selection) with a TRANSITION_DURATION constant in MyApplication.