Update Counter Multiplier Button Text When IndividualHunt Finishes
Previously, when modifying the counter multiplier value  in IndividualHunt and navigating back to MainActivity, the counter multiplier button in MainActivity wasn't being updated to reflect the new value of counterMultiplier. Added a line of code to MainActivity's 'onResume()' function to reset the text of the counter multiplier button when returning from the IndividualHunt page.