Added Nickname and Location to Shiny Hunts + Set Max Counter and Phase Values in Filter Menu
Upgraded the database and ShinyHuntModel to support "nickname" and "location" fields. Refactored all functions that interface with the database's Shiny Hunt table to support these new fields, and made the fields accessible to view and modify in the Individual Hunt page.

Applied max value filters to the Filter Menu's counter and phase filters to prevent integer overflow from occurring.