IndividualHunt Background Transition Animation + Adjusted Colors for Gold Gradient
Added a transition animation for the background of the Individual Hunt page when the hunt completion checkbox is checked or unchecked.

Also slightly adjusted the gold gradient (for complete hunts). Made the starting color lighter, and shifted the darker color to be closer to brown/orange than to green as it previously was.