Adjustments to Individual Hunt Layout, Textbox Drawable, Text Sizes, and Button Margins
Slightly reduced text size and button margins to make the UI of the Individual Hunt page fit on smaller phones (previously, the "Nickname:" label automatically created a new line due to being slightly too long, and the increment counter/phase buttons were clipping out of the parent layout). Also created a 'label_top_margin' value to replace the hardcoded top margins of each label, and reduced the top margin from 50dp to 40dp.

Slightly increased the transparency of the 'ui_textbox' drawable, and applied it to the background of all EditText fields in the Individual Hunt page. Created a 'textbox_padding' value and used it to replace hardcoded paddings for all EditText fields in the Individual Hunt page and Filter menu.