Fixed onUpgrade() function + Renamed all database queries
Previously, the onUpgrade() function in DBHelper would entirely drop and recreate the database, clearing all user data (aside from dev mock data if enabled). Now, it checks if the old version of the database is less than 7 (6 was the official release version, and 7 is currently the newest version that adds pokeballs to the database). If the old version is less than 7, the app now creates the Pokeball table and copies the user's shiny hunt data into a new ShinyHunt table containing a foreign key to the Pokeball table. It also recreates and repopulates the PokemonForm and Game tables to update the resource IDs of the images.

Also renamed all database queries for better clarity.