Data Transfer Page Created
In the navigation menu, replaced the previous "Export Hunts" and "Import Hunts" buttons with a single "Data Transfer button" which takes the user to a new Data Transfer page containing buttons to export and import shiny hunt logs, along with explanations of how data exporting and importing will work. The actual export and import functionality has not been implemented yet.